Milestone/p2---ftds-018-rmt---m2-andik-alfauzi/deployment/requirements.txt
pandas
seaborn
matplotlib
Pillow
numpy
scikit-learn
ensemble
tensorflow-cpu
keras
streamlit
waitress